refactor: remove sha.js in favour of nobel hashes
