+++ package.json
@@ -129,6 +129,7 @@
     "@emotion/react": "11.7.1",
     "@emotion/styled": "11.6.0",
     "@ledgerhq/hw-transport-webusb": "6.24.1",
+    "@noble/hashes": "1.1.4",
     "@noble/secp256k1": "1.6.3",
     "@reach/alert": "0.15.3",
     "@reach/auto-id": "0.15.3",
@@ -178,7 +179,7 @@
     "ecdsa-sig-formatter": "1.0.11",
     "formik": "2.2.9",
     "jotai": "1.5.3",
-    "jsontokens": "3.0.0",
+    "jsontokens": "4.0.1",
     "limiter": "2.1.0",
     "lodash.get": "4.4.2",
     "mdi-react": "7.5.0",
@@ -199,7 +200,6 @@
     "react-virtuoso": "2.19.1",
     "redux-persist": "6.0.0",
     "rxjs": "7.5.7",
-    "sha.js": "2.4.11",
     "ts-debounce": "4",
     "use-events": "1.4.2",
     "use-latest": "1.2.0",
