+++ src/app/features/ledger/flows/message-signing/message-signing.utils.ts
@@ -1,0 +1,36 @@
+import { sha256 } from '@noble/hashes/sha256';
+import { bytesToHex } from '@stacks/common';
+import {
+  ChainID,
+  ClarityType,
+  ClarityValue,
+  cvToString,
+  encodeStructuredData,
+} from '@stacks/transactions';
+
+import { SignedMessageStructured } from '@shared/signature/signature-types';
+
+import { whenStxChainId } from '@app/common/utils';
+
+export function cvToDisplay(cv: ClarityValue): string {
+  return cvToString(cv).replaceAll('"', '');
+}
+
+export function chainIdToDisplay(chainIdCv: ClarityValue): string {
+  if (chainIdCv.type !== ClarityType.UInt) return '';
+  const chainIdString = cvToString(chainIdCv);
+  const chainId = parseInt(chainIdString.replace('u', ''));
+  if (!Object.values(ChainID).includes(chainId)) return '';
+
+  return whenStxChainId(chainId as ChainID)({
+    [ChainID.Testnet]: 'Testnet',
+    [ChainID.Mainnet]: 'Mainnet',
+  });
+}
+
+export function deriveStructuredMessageHash({
+  domain,
+  message,
+}: Omit<SignedMessageStructured, 'messageType'>) {
+  return bytesToHex(sha256(encodeStructuredData({ message, domain })));
+}
