+++ src/app/pages/message-signing-request/components/structured-data-box.tsx
@@ -9,7 +9,8 @@ import {
   chainIdToDisplay,
   cvToDisplay,
   deriveStructuredMessageHash,
-} from '../message-signing.utils';
+} from '@app/features/ledger/flows/message-signing/message-signing.utils';
+
 import { ClarityValueListDisplayer } from './clarity-value-list';
 import { HashDrawer } from './hash-drawer';
